feat: work on eduction and experience controller, CRUD done

+++ app/Http/Controllers/API/EducationController.php
@@ -5,18 +5,22 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Models\Education;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class EducationController extends Controller
 {
-    public function index()
+    public function getUserEductions()
     {
-        $educations = Education::all();
-        return response()->json($educations);
+        $educations = auth()->user()->educations;
+        return response()->json([
+            'success' => true,
+            'data' => $educations,
+        ]);
     }
 
     public function store(Request $request)
     {
-        $request->validate([
+        $validators = Validator::make($request->all(), [
             'school' => 'required|string',
             'degree' => 'required|string',
             'field_of_study' => 'required|string',
@@ -24,6 +28,14 @@ class EducationController extends Controller
             'end_date' => 'required|date',
             'resume_id' => 'required|integer',
         ]);
+
+        if($validators->fails()){
+            return response()->json([
+                'success' => false,
+                'message' => $validators->errors()->all(),
+            ]);
+        }
+
         $education = new Education([
             'school' => $request->school,
             'degree' => $request->degree,
@@ -34,6 +46,7 @@ class EducationController extends Controller
         ]);
         $education->save();
         return response()->json([
+            'success' => true,
             'message' => 'Successfully created education!',
         ], 201);
     }
@@ -46,7 +59,7 @@ class EducationController extends Controller
 
     public function update(Request $request, $id)
     {
-        $request->validate([
+        $validators = Validator::make($request->all(), [
             'school' => 'required|string',
             'degree' => 'required|string',
             'field_of_study' => 'required|string',
@@ -54,19 +67,68 @@ class EducationController extends Controller
             'end_date' => 'required|date',
             'resume_id' => 'required|integer',
         ]);
+
+        if ($validators->fails()) {
+            return response()->json([
+                'success' => false,
+                'message' => $validators->errors()->all(),
+            ]);
+        }
         $education = Education::find($id);
-        $education->update($request->all());
-        return response()->json([
-            'message' => 'Successfully updated education!',
-        ], 201);
+
+        if (!$education) {
+            return response()->json([
+                'success' => false,
+                'message' => 'no education found with id ' . $id,
+            ]);
+        }
+        try {
+            $education->update($request->all());
+            return response()->json([
+                'success' => true,
+                'message' => 'Successfully updated education!',
+            ], 201);
+
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => $e->getMessage(),
+            ]);
+
+        }
+
     }
 
     public function destroy($id)
     {
         $education = Education::find($id);
+        if (!$education) {
+            return response()->json([
+                'success' => false,
+                'message' => 'no educiton found with id ' . $id
+            ]);
+        }
         $education->delete();
         return response()->json([
             'message' => 'Successfully deleted education!',
         ], 201);
     }
+
+    public function getResumeEductions($resumeId)
+    {
+        $userEductions = auth()->user()->resumes()->education::where('resume_id', $resumeId)->get();
+        if(!$userEductions){
+            return response()->json([
+                'success' => false,
+                'message' => 'no eductions found for this resume',
+                'data' => $userEductions,
+            ]);
+        }
+
+        return response()->json([
+            'success' => true,
+            'data' => $userEductions
+        ]);
+    }
+
 }
